Add Spring Security (#3)

+++ src/main/java/com/lukaskucera/numberneighbors/controller/GameController.java
@@ -1,14 +1,21 @@
 package com.lukaskucera.numberneighbors.controller;
 
 import com.lukaskucera.numberneighbors.entity.Game;
+import com.lukaskucera.numberneighbors.entity.Player;
+import com.lukaskucera.numberneighbors.request.NewGameRequest;
+import com.lukaskucera.numberneighbors.response.NewGameResponse;
 import com.lukaskucera.numberneighbors.service.GameService;
 import com.lukaskucera.numberneighbors.service.GameServiceImpl;
+import com.lukaskucera.numberneighbors.service.JwtService;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,23 +23,35 @@ import org.springframework.web.bind.annotation.RestController;
 public class GameController {
   private final GameService gameService;
 
-  public GameController(GameServiceImpl gameService) {
+  private final JwtService jwtService;
+
+  public GameController(GameServiceImpl gameService, JwtService jwtService) {
     this.gameService = gameService;
+    this.jwtService = jwtService;
   }
 
   @PostMapping(value = "/games")
-  public ResponseEntity<Game> newGame() {
-    return ResponseEntity.ok(gameService.newGame());
+  public ResponseEntity<NewGameResponse> newGame(
+      @RequestBody @NotNull NewGameRequest newGameRequest) {
+    final Game game = gameService.newGame(newGameRequest.hostName());
+    final Player hostPlayer = game.getHost();
+    final String token =
+        jwtService.generatePlayerToken(hostPlayer.getName(), hostPlayer.getId(), game.getId());
+
+    return ResponseEntity.ok(new NewGameResponse(game, token));
   }
 
   @GetMapping(value = "/games/{id}")
-  public ResponseEntity<Game> game(@PathVariable Long id) {
+  public ResponseEntity<Game> game(
+      @PathVariable Long id, @NotNull JwtAuthenticationToken jwtToken) {
+    gameService.checkGameAccess(id, jwtToken);
     return ResponseEntity.ok(gameService.getGameById(id));
   }
 
   @DeleteMapping(value = "/games/{id}")
   @ResponseStatus(value = HttpStatus.NO_CONTENT)
-  public void deleteGame(@PathVariable Long id) {
+  public void deleteGame(@PathVariable Long id, @NotNull JwtAuthenticationToken jwtToken) {
+    gameService.checkGameAccess(id, jwtToken);
     gameService.deleteGameById(id);
   }
 }
